feature: std as a default feature (#133)
