+++ rust/Cargo.toml
@@ -18,6 +18,10 @@ num = "0.2.0"
 num-bigint = "0.2"
 ripemd160 = "0.8.0"
 sha2 = "0.8.0"
-serde = { version = "1.0.99", features = ["derive"] }
-serde_json = "1.0"
-hex = "0.4.0"
+hex = "0.4.2"
+serde_json = { version = "1.0", optional = true }
+serde = { version = "1.0.99", features = ["derive"], optional = true }
+
+[features]
+default=["std"]
+std=["num/std", "num-bigint/std", "serde", "serde_json", "hex/std", "sha2/std", "ripemd160/std"]
